requirements.txt
bz2file==0.98
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.2
python-decouple==3.6
python-dotenv==0.21.1
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
threadpoolctl==3.1.0
urllib3==1.26.14
Werkzeug==2.2.2
